myproject/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.9
django-debug-toolbar==4.3.0
django-embed-video==1.4.9
djangorestframework==3.14.0
flake8==7.0.0
fontawesomefree==6.5.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isort==5.13.2
mccabe==0.7.0
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pyOpenSSL==24.0.0
pytz==2024.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==2.2.1
zope.interface==6.2
